% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Title Function to Get Weather Data for a given city}
\usage{
get_data(input_city, forecast_days)
}
\arguments{
\item{input_city}{The name of the city that weather needs to be fetched}

\item{forecast_days}{The number of days that need to be forecasted}
}
\value{
current_weather_data
}
\description{
Title Function to Get Weather Data for a given city
}
\examples{
get_data("Bangalore", 1)

}
